% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{annotate_tree}
\alias{annotate_tree}
\title{Annotate tree}
\usage{
annotate_tree(
  tree,
  annot_dt,
  value_col = "count",
  on = c(code = "code", code_type = "code_type"),
  annot_name = "value",
  no_match = 0
)
}
\arguments{
\item{tree}{a tree}

\item{annot_dt}{a data.table, data to use for annotation}

\item{value_col}{a string, column name in `annot_dt` with the annotation values.}

\item{on}{a named character vector, how to join the data e.g. c("annot_dt_col" = "node$data$name"), can have multiple conditions e.g. c(code=code, code_type=code_type)}

\item{annot_name}{new attribute name (inserted into node$data)}

\item{no_match}{any value, what to annotate with if there is no join data}
}
\value{
a nested list, a tree
}
\description{
Annotate tree
}
