% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biovu_counts}
\alias{biovu_counts}
\title{BioVu count data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 136949 rows and 3 columns.
}
\usage{
biovu_counts
}
\description{
BioVu count data
}
\keyword{datasets}
