% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_id}
\alias{clean_id}
\title{Clean ID string}
\usage{
clean_id(str)
}
\arguments{
\item{str}{a string to clean}
}
\value{
a clean string
}
\description{
id cleaning (only allowed letters, numbers, underscores)
}
