% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{code_from_label}
\alias{code_from_label}
\title{Extract code from tree label}
\usage{
code_from_label(label)
}
\arguments{
\item{label}{string label}
}
\value{
a string, the code
}
\description{
Extract code from tree label
}
