% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeconsensus-package.R
\docType{package}
\name{codeconsensus-package}
\alias{codeconsensus}
\alias{codeconsensus-package}
\title{codeconsensus: Electronic Health Records Phenotyping}
\description{
Enabling disease phenotyping by consensus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicksunderland/codeconsensus}
  \item Report bugs at \url{https://github.com/nicksunderland/codeconsensus/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Sunderland \email{nicholas.sunderland@bristol.ac.uk}

}
\keyword{internal}
