% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{cpt4_tree}
\alias{cpt4_tree}
\title{CPT4 tree}
\usage{
cpt4_tree(cpt4_dt, concept_id, regex = ".")
}
\arguments{
\item{cpt4_dt}{a data.table of cpt4 codes with columns CODE and DESCRIPTION}

\item{regex}{regular expression to search for}
}
\value{
a nested list
}
\description{
CPT4 tree
}
\references{
https://www.cdc.gov/nhsn/xls/cpt-pcm-nhsn.xlsx
}
