% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfphenotyping-package.R
\docType{package}
\name{hfphenotyping-package}
\alias{hfphenotyping}
\alias{hfphenotyping-package}
\title{hfphenotyping: Heart Failure Phenotyping}
\description{
Enabling heart failure phenotyping by consensus.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nicksunderland/hfphenotyping}
  \item Report bugs at \url{https://github.com/nicksunderland/hfphenotyping/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Sunderland \email{nicholas.sunderland@bristol.ac.uk}

}
\keyword{internal}
