% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{icd10_tree}
\alias{icd10_tree}
\title{ICD10 tree}
\usage{
icd10_tree(icd10, regex = ".")
}
\arguments{
\item{icd10}{list output from xml2::read_xml -> xml2::as_list (https://www.cms.gov/files/zip/2024-code-tables-tabular-and-index-updated-02/01/2024.zip)}

\item{regex}{regular expression to search for}
}
\value{
a nested list
}
\description{
ICD10 tree
}
