% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{icd9_procedure_tree}
\alias{icd9_procedure_tree}
\title{ICD9 procedure tree}
\usage{
icd9_procedure_tree(icd9_proc_dt, concept_id, regex = ".")
}
\arguments{
\item{icd9_proc_dt}{a data.table of ICD9 procedure codes with columns CODE and DESCRIPTION}

\item{regex}{regular expression to search for}
}
\value{
a nested list
}
\description{
ICD9 procedure tree
}
