% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{icd9_tree}
\alias{icd9_tree}
\title{ICD9 tree}
\usage{
icd9_tree(icd9_dt, concept_id, regex = ".")
}
\arguments{
\item{icd9_dt}{a data.table of ICD9 codes with columns CODE and DESCRIPTION}

\item{regex}{regular expression to search for}
}
\value{
a nested list
}
\description{
ICD9 tree
}
