% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{make_icd11_tree}
\alias{make_icd11_tree}
\title{make_icd11_tree}
\usage{
make_icd11_tree(icd11_dt, concept_id, regex = ".", level = 1, prepocess = TRUE)
}
\arguments{
\item{icd11_dt}{the dt read from SimpleTabulation-ICD-11-MMS-en.txt}

\item{concept_id}{the concept id}

\item{regex}{the regular expression to search for}

\item{level}{internal use for recursion}

\item{prepocess}{internal use, dont change}
}
\value{
a nested tree
}
\description{
make_icd11_tree
}
\references{
https://icd.who.int/browse/2024-01/mms/en download file from the info tab
}
