% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhs_counts}
\alias{nhs_counts}
\title{NHS count data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 173372 rows and 6 columns.
}
\usage{
nhs_counts
}
\description{
NHS count data
}
\keyword{datasets}
