% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_database.R
\name{query_db}
\alias{query_db}
\title{database connection and query function}
\usage{
query_db(query_str = NULL, type = "get", ...)
}
\arguments{
\item{query_str}{string, valid SQL (or name of dt if 'write' option)}

\item{type}{string, either get or send}

\item{...}{arguments to pass to DBI::dbWriteTable}
}
\value{
a data.table
}
\description{
database connection and query function
}
