% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{snomed_tree}
\alias{snomed_tree}
\title{SNOMED tree}
\usage{
snomed_tree(hierarch_codes, concept_id)
}
\arguments{
\item{hierarch_codes}{output from Rdiagnosislist::showCodelistHierarchy}
}
\value{
a nested list
}
\description{
SNOMED tree
}
