% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tree_attributes}
\alias{tree_attributes}
\title{Get tree attributes}
\usage{
tree_attributes(tree, attribute_name, result = list())
}
\arguments{
\item{tree}{a js tree object}

\item{attribute_name}{str, the attribute name (label, code, code_type, description, stselected, stopened)}

\item{result}{for recursive use}
}
\value{
a list of name = attribute
}
\description{
Get tree attributes
}
