% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{TreeElement}
\alias{TreeElement}
\title{TreeElement S3 class}
\usage{
TreeElement(
  name,
  db_name,
  stselected = FALSE,
  stopened = TRUE,
  children = list()
)
}
\arguments{
\item{name}{the label}

\item{db_name}{the database column name}

\item{stselected}{logical}

\item{stopened}{logical}

\item{data}{empty character ""}
}
\value{
a TreeElement S3 object
}
\description{
TreeElement S3 class
}
