% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeNode.R
\name{TreeNode}
\alias{TreeNode}
\title{JS TreeNode S3 class}
\usage{
TreeNode(
  text,
  type = "code",
  data = list(),
  checked = FALSE,
  selected = FALSE,
  opened = TRUE,
  disabled = FALSE,
  children = list()
)
}
\arguments{
\item{text}{string, the tree label}

\item{type}{string, a valid NodeTypes (see above)}

\item{data}{list, any required data attributes}

\item{checked}{logical}

\item{selected}{logical}

\item{opened}{logical}

\item{disabled}{logical}

\item{children}{empty list (to be filled with other TreeNodes if needed)}
}
\value{
a TreeNode S3 object
}
\description{
JS TreeNode S3 class
}
