% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ukbb_counts}
\alias{ukbb_counts}
\title{NHS count data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20664 rows and 5 columns.
}
\usage{
ukbb_counts
}
\description{
NHS count data
}
\keyword{datasets}
