% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_id}
\alias{clean_id}
\title{Clean ID string}
\usage{
clean_id(str, check = FALSE, to_lower = TRUE)
}
\arguments{
\item{str}{a string to clean}

\item{check}{a logical, whether to use as a checking function}

\item{to_lower}{logical, whether to convert to lowercase}
}
\value{
a clean string
}
\description{
id cleaning (only allowed letters, numbers, underscores)
}
