% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{opcs_tree}
\alias{opcs_tree}
\title{OPCS tree}
\usage{
opcs_tree(opcs_dt, concept_id, regex = ".")
}
\arguments{
\item{opcs_dt}{a data.table of OPCS codes with columns CODE and DESCRIPTION}

\item{regex}{regular expression to search for}
}
\value{
a nested list
}
\description{
OPCS tree
}
