% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{query_db}
\alias{query_db}
\title{database connection and query function}
\usage{
query_db(query_str = NULL, type = "get", table = NULL, value = NULL)
}
\arguments{
\item{query_str}{string, valid SQL (or name of dt if 'write' option)}

\item{type}{string, either get or send}

\item{table}{database table name (for read and write options)}

\item{value}{values (for update and write options)}
}
\value{
a data.table
}
\description{
database connection and query function
}
