% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{snomed_tree}
\alias{snomed_tree}
\title{SNOMED tree}
\usage{
snomed_tree(hierarch_codes)
}
\arguments{
\item{hierarch_codes}{output from Rdiagnosislist::showCodelistHierarchy}

\item{row_id}{NA for recursive use}
}
\value{
a nested list
}
\description{
SNOMED tree
}
