% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{TreeElement}
\alias{TreeElement}
\title{TreeElement S3 class}
\usage{
TreeElement(
  code,
  code_type,
  description,
  stselected = FALSE,
  stopened = TRUE,
  children = list()
)
}
\arguments{
\item{code}{string, the code e.g. I510}

\item{code_type}{string, the code tyoe e.g. "ICD10"}

\item{description}{string, the code's description}

\item{stselected}{logical}

\item{stopened}{logical}

\item{children}{empty list for recursive use (will potentially be filled with other TreeElements)}
}
\value{
a TreeElement S3 object
}
\description{
TreeElement S3 class
}
